move router out of interface.py
